added InMemoryEventStore for unit testing pbjx handlers and event subscribers.
Unit tests need to be able to issue commands, subscribe to events, etc. and then assert
that the event store was given events (command handler calling ->putEvents).

It also allows for putting events into the store to assert conditions where handlers
expect that data is in the store (or not) already.